Adds an explicit requirement of the 'set' dependency.
The lack of "require 'set'" was not a problem when innertube was used in rails
or tested with RSpec, because 'set' is already loaded by those frameworks by the
time innertube can be initialized. It was only a problem in Ruby projects that
didn't happen to already load 'set' before using innertube.