add full ASL 2.0 text to license
All derivative copies of innertube are required to contain the full text
of the ASL 2.0 license (as described in the sentence "You must give any
other recipients of the Work or Derivative Works a copy of this
License"). In order to ease development and distribution of innertube,
it makes sense to simply include the full text in the upstream project.